lab 12/requirements.txt
flask==3.0.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy>=1.24
pandas>=2.0